Make the NoopRedis fixture wait for the server to come up
The `redis_nooproc` was a NOOP named tuple. The fixture is usually used
when users control their own redis instance. It would be easy to assume,
that the fixture would wait (block) for the server to be ready to
connect to. This was not the case.

Change the data structure to a class inheriting from `TCPExecutor`. The
start up routine now wait's for the fixture to create a successful
socket connection before the timout is met, otherwise fail.

Note: This only uses `AF_INET` sockets (no IPv6 support).

Fixes https://github.com/ClearcodeHQ/pytest-redis/issues/388